refactor(workers-dev): move the stack picker into tui/stacks.rs
